Return a taxonomic table matching the requested ids

Return a taxonomic table matching the requested ids

by_id(id, provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd", "ott", "iucn"), type = c("taxonID", "acceptedNameUsageID"),
  collect = TRUE, db = td_connect())

Arguments

id
taxonomic id, in prefix format

provider
from which provider should the hierarchy be returned? Default is 'itis'.

type
id type. Can be taxonID or acceptedNameUsageID, see details.

collect
logical, default TRUE. Should we return an in-memory data.frame (default, usually the most convenient), or a reference to lazy-eval table on disk (useful for very large tables on which we may first perform subsequent filtering operations.)

db
a connection to the taxadb database. See details.

Value

a data.frame with id and name of all matching species

Details

Use type="acceptedNameUsageID" to return all rows for which this ID is the accepted ID, including both synonyms and and accepted names (since both all synonyms of a name share the same acceptedNameUsageID.) Use taxonID (default) to only return those rows for which the Scientific name corresponds to the taxonID.

Some providers (e.g. ITIS) assign taxonIDs to synonyms, most others only assign IDs to accepted names. In the latter case, this means requesting taxonID will only match accepted names, while requesting matches to the acceptedNameUsageID will also return any known synonyms. See examples.

See also

Examples

# \donttest{
  # \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  # }

by_id(c("ITIS:1077358", "ITIS:175089"))

#> Native bulk importer found, attempting fast import of dwc_itis.tsv.bz2

#> Native import failed, falling back on R-based parser

#> Importing dwc_itis.tsv.bz2 in 100000 line chunks:

#> 	...Done! (in 58.46806 secs)

#> # A tibble: 2 x 17
#>   input  sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
#>   <chr> <int> <chr>   <chr>          <chr>     <chr>            <chr>          
#> 1 ITIS…     1 ITIS:1… Antrostomus v… species   ITIS:1077358     accepted       
#> 2 ITIS…     2 ITIS:1… Anas cyanopte… species   ITIS:175089      accepted       
#> # … with 10 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>

by_id("ITIS:1077358", type="acceptedNameUsageID")

#> # A tibble: 3 x 17
#>   input  sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
#>   <chr> <int> <chr>   <chr>          <chr>     <chr>            <chr>          
#> 1 ITIS…     1 ITIS:1… Caprimulgus v… species   ITIS:1077358     synonym        
#> 2 ITIS…     1 ITIS:1… Caprimulgus v… subspeci… ITIS:1077358     synonym        
#> 3 ITIS…     1 ITIS:1… Antrostomus v… species   ITIS:1077358     accepted       
#> # … with 10 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>

# }