Return the full classification hierarchy for requested species or ids

Return the full classification hierarchy for requested species or ids

classification(species = NULL, id = NULL, authority = c("itis",
  "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd"), collect = TRUE,
  db = td_connect())

Arguments

species
a character vector of species names, typically specified as (Genus species or Genus species epithet)

id
alternately users can provide a vector of species ids. IDs must be prefixed matching the requested authority. See id column returned by most taxadb functions for examples.

authority
from which authority should the hierachy be returned? Default is 'itis'.

collect
logical, default TRUE. Should we return an in-memory data.frame (default, usually the most convenient), or a reference to lazy-eval table on disk (useful for very large tables on which we may first perform subsequent filtering operations.)

db
a connection to the taxadb database. See details.

Value

a data.frame with one row for each requested species, giving the species id, and column for each of the unique-rank levels of the species. Note that the different authorities recognize a range of different ranks.

Details

Some authorities recognize multiple values of the same rank level, (i.e. a species may be assigned to two different "suborders"). In this case, only one has been included in the "hierarchy" schema. Likewise, some authorities (i.e. NCBI) use some unnamed rank levels (i.e. a scientific name is associated with a clade, but the clade is not associated with any traditionally recognized rank). In these cases, see the "long" schema for more complete classification.