get_names

Translate identifiers into scientific names

get_names(id, db = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd", "ott", "iucn"), format = c("guess", "prefix", "bare", "uri"),
  taxadb_db = td_connect())

Arguments

id
a list of taxonomic identifiers.

db
abbreviation code for the provider. See details.

format
Format for the input identifier, one of

guess Will determine the prefix automatically, but will be slower than if you specify this in advance!

prefix (e.g. NCBI:9606), Preferred (fastest) format.

bare (e.g. 9606), (But must mach provider db!)

uri (e.g. http://ncbi.nlm.nih.gov/taxonomy/9606).

taxadb_db
Connection to from [td_connect]().

Value

a vector of names, of the same length as the input ids. Any unmatched IDs will return as NAs.

Details

Like all taxadb functions, this function will run fastest if a local copy of the provider is installed in advance using [td_create()].

See also

Examples

# \donttest{
get_names(180092)

#> [1] "Homo sapiens"

get_names(c("ITIS:180092", "ITIS:179913"))

#> [1] "Homo sapiens" "Mammalia"    

get_names(c("ITIS:180092", "ITIS:179913"), format = "prefix")

#> [1] "Homo sapiens" "Mammalia"    

# }