Creates a data frame with column name given by by, and values given by the vector x, and then uses this table to do a filtering join, joining on the by column to return all rows matching the x values (scientificNames, taxonIDs, etc).

Creates a data frame with column name given by by, and values given by the vector x, and then uses this table to do a filtering join, joining on the by column to return all rows matching the x values (scientificNames, taxonIDs, etc).

filter_by(x, by, provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb",
  "slb", "wd", "ott", "iucn"), schema = c("dwc", "common"),
  collect = TRUE, db = td_connect(), ignore_case = TRUE)

Arguments

x
a vector of values to filter on

by
a column name in the taxa_tbl (following Darwin Core Schema terms). The filtering join is executed with this column as the joining variable.

provider
from which provider should the hierarchy be returned? Default is 'itis'.

schema
One of "dwc" (for Darwin Core data) or "common" (for the Common names table.)

collect
logical, default TRUE. Should we return an in-memory data.frame (default, usually the most convenient), or a reference to lazy-eval table on disk (useful for very large tables on which we may first perform subsequent filtering operations.)

db
a connection to the taxadb database. See details.

ignore_case
should we ignore case (capitalization) in matching names? default is TRUE.

Value

a data.frame in the Darwin Core tabular format containing the matching taxonomic entities.

See also

Examples

# \donttest{
  # \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  # }

sp <- c("Trochalopteron henrici gucenense",
        "Trochalopteron elliotii")
filter_by(sp, "scientificName")

#> # A tibble: 3 x 17
#>    sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
#>   <int> <chr>   <chr>          <chr>     <chr>            <chr>          
#> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
#> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
#> 3     2 ITIS:9… Trochaloptero… species   ITIS:916116      accepted       
#> # … with 11 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>

filter_by(c("ITIS:1077358", "ITIS:175089"), "taxonID")

#> # A tibble: 2 x 17
#>    sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
#>   <int> <chr>   <chr>          <chr>     <chr>            <chr>          
#> 1     1 ITIS:1… Antrostomus v… species   ITIS:1077358     accepted       
#> 2     2 ITIS:1… Anas cyanopte… species   ITIS:175089      accepted       
#> # … with 11 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>

filter_by("Aves", "class")

#> # A tibble: 12,704 x 17
#>     sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
#>    <int> <chr>   <chr>          <chr>     <chr>            <chr>          
#>  1     1 ITIS:1… Rollandia mic… species   ITIS:562791      synonym        
#>  2     1 ITIS:1… Ardea helias   species   ITIS:176409      synonym        
#>  3     1 ITIS:1… Mesites varie… species   ITIS:707807      synonym        
#>  4     1 ITIS:1… Mesites unico… species   ITIS:561100      synonym        
#>  5     1 ITIS:1… Mesoenas unic… species   ITIS:561100      synonym        
#>  6     1 ITIS:1… Mesoenas vari… species   ITIS:707807      synonym        
#>  7     1 ITIS:1… Podiceps micr… species   ITIS:562791      synonym        
#>  8     1 ITIS:1… Colymbus rufi… species   ITIS:563259      synonym        
#>  9     1 ITIS:1… Podiceps tric… species   ITIS:824133      synonym        
#> 10     1 ITIS:1… Colymbus domi… species   ITIS:174508      synonym        
#> # … with 12,694 more rows, and 11 more variables: update_date <chr>,
#> #   kingdom <chr>, phylum <chr>, class <chr>, order <chr>, family <chr>,
#> #   genus <chr>, specificEpithet <chr>, vernacularName <chr>,
#> #   infraspecificEpithet <chr>, input <chr>

# }