Creates a data frame with column name given by by, and values given by the vector x, and then uses this table to do a filtering join, joining on the by column to return all rows matching the x values (scientificNames, taxonIDs, etc).

Creates a data frame with column name given by by, and values given by the vector x, and then uses this table to do a filtering join, joining on the by column to return all rows matching the x values (scientificNames, taxonIDs, etc).

filter_by(x, by, provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb",
  "slb", "wd", "ott", "iucn"), collect = TRUE, db = td_connect(),
  ignore_case = TRUE)

Arguments

x
a vector of values to filter on

by
a column name in the taxa_tbl (following Darwin Core Schema terms). The filtering join is executed with this column as the joining variable.

provider
from which provider should the hierachy be returned? Default is 'itis'.

collect
logical, default TRUE. Should we return an in-memory data.frame (default, usually the most convenient), or a reference to lazy-eval table on disk (useful for very large tables on which we may first perform subsequent filtering operations.)

db
a connection to the taxadb database. See details.

ignore_case
should we ignore case (capitalization) in matching names? default is TRUE.

Value

a data.frame in the Darwin Core tabular format containing the matching taxonomic entities.

See also

Examples

     ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())

sp <- c("Trochalopteron henrici gucenense",
        "Trochalopteron elliotii")
filter_by(sp, "scientificName")

#> # A tibble: 3 x 17
#>   taxonID acceptedNameUsa… update_date scientificName taxonRank taxonomicStatus
#>   <chr>   <chr>            <chr>       <chr>          <chr>     <chr>          
#> 1 ITIS:9… ITIS:916116      2013-11-04  Trochaloptero… subspeci… synonym        
#> 2 ITIS:9… ITIS:916117      2013-11-04  Trochaloptero… subspeci… synonym        
#> 3 ITIS:9… ITIS:916116      2013-11-04  Trochaloptero… species   accepted       
#> # … with 11 more variables: kingdom <chr>, phylum <chr>, class <chr>,
#> #   order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>, sort <int>

filter_by(c("ITIS:1077358", "ITIS:175089"), "taxonID")

#> # A tibble: 2 x 17
#>   taxonID acceptedNameUsa… update_date scientificName taxonRank taxonomicStatus
#>   <chr>   <chr>            <chr>       <chr>          <chr>     <chr>          
#> 1 ITIS:1… ITIS:1077358     2018-06-26  Antrostomus v… species   accepted       
#> 2 ITIS:1… ITIS:175089      2007-01-23  Anas cyanopte… species   accepted       
#> # … with 11 more variables: kingdom <chr>, phylum <chr>, class <chr>,
#> #   order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>, sort <int>

filter_by("Aves", "class")

#> # A tibble: 12,614 x 17
#>    taxonID acceptedNameUsa… update_date scientificName taxonRank taxonomicStatus
#>    <chr>   <chr>            <chr>       <chr>          <chr>     <chr>          
#>  1 ITIS:1… ITIS:562791      2016-04-29  Rollandia mic… species   synonym        
#>  2 ITIS:1… ITIS:176409      2016-04-29  Ardea helias   species   synonym        
#>  3 ITIS:1… ITIS:707807      2016-04-29  Mesites varie… species   synonym        
#>  4 ITIS:1… ITIS:561100      2016-04-29  Mesites unico… species   synonym        
#>  5 ITIS:1… ITIS:561100      2016-04-29  Mesoenas unic… species   synonym        
#>  6 ITIS:1… ITIS:707807      2016-04-29  Mesoenas vari… species   synonym        
#>  7 ITIS:1… ITIS:562791      2016-04-29  Podiceps micr… species   synonym        
#>  8 ITIS:1… ITIS:563259      2016-04-29  Colymbus rufi… species   synonym        
#>  9 ITIS:1… ITIS:824133      2016-04-29  Podiceps tric… species   synonym        
#> 10 ITIS:1… ITIS:174508      2016-04-29  Colymbus domi… species   synonym        
#> # … with 12,604 more rows, and 11 more variables: kingdom <chr>, phylum <chr>,
#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>, sort <int>