Match names that start or contain a specified text string

Match names that start or contain a specified text string

fuzzy_filter(
  name,
  by = c("scientificName", "vernacularName"),
  provider = getOption("taxadb_default_provider", "itis"),
  match = c("contains", "starts_with"),
  version = latest_version(),
  db = td_connect(),
  ignore_case = TRUE,
  collect = TRUE
)

Arguments

name
vector of names (scientific or common, see by) to be matched against.

by
a column name in the taxa_tbl (following Darwin Core Schema terms). The filtering join is executed with this column as the joining variable.

provider
from which provider should the hierarchy be returned? Default is 'itis', which can also be configured using options(default_taxadb_provider=..."). See [td_create] for a list of recognized providers.

match
should we match by names starting with the term or containing the term anywhere in the name?

version
Which version of the taxadb provider database should we use? defaults to latest. See tl_import for details.

db
a connection to the taxadb database. See details.

ignore_case
should we ignore case (capitalization) in matching names? default is TRUE.

collect
logical, default TRUE. Should we return an in-memory data.frame (default, usually the most convenient), or a reference to lazy-eval table on disk (useful for very large tables on which we may first perform subsequent filtering operations.)

Details

Note that fuzzy filter will be fast with an single or small number of names, but will be slower if given a very large vector of names to match, as unlike other filter_ commands, fuzzy matching requires separate SQL calls for each name. As fuzzy matches should all be confirmed manually in any event, e.g. not every common name containing "monkey" belongs to a primate species.

This method utilizes the database operation %like% to filter tables without loading into memory. Note that this does not support the use of regular expressions at this time.

Examples

# \donttest{
  # \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
   options("taxadb_default_provider"="itis_test")
  # }

## match any common name containing:
name <- c("woodpecker", "monkey")
fuzzy_filter(name, "vernacularName")

#> # A tibble: 418 x 16
#>    vernacularName  language update_date acceptedNameUsag… taxonID scientificName
#>    <chr>           <chr>    <chr>       <chr>             <chr>   <chr>         
#>  1 South American… English  2002-07-30  ITIS:180095       ITIS:9… Simia apedia  
#>  2 South American… English  2002-07-30  ITIS:180095       ITIS:9… Lemur leucops…
#>  3 South American… English  2002-07-30  ITIS:180095       ITIS:9… Simia morta   
#>  4 South American… English  2002-07-30  ITIS:180095       ITIS:9… Chrysothrix n…
#>  5 South American… English  2002-07-30  ITIS:180095       ITIS:9… Sapajou saimi…
#>  6 South American… English  2002-07-30  ITIS:180095       ITIS:9… Saimiri sciur…
#>  7 South American… English  2002-07-30  ITIS:180095       ITIS:1… Saimiri sciur…
#>  8 Common Squirre… English  2014-06-26  ITIS:180095       ITIS:9… Simia apedia  
#>  9 Common Squirre… English  2014-06-26  ITIS:180095       ITIS:9… Lemur leucops…
#> 10 Common Squirre… English  2014-06-26  ITIS:180095       ITIS:9… Simia morta   
#> # … with 408 more rows, and 10 more variables: taxonRank <chr>,
#> #   taxonomicStatus <chr>, kingdom <chr>, phylum <chr>, class <chr>,
#> #   order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
#> #   infraspecificEpithet <chr>

## match scientific name
fuzzy_filter("Chera", "scientificName",
             match = "starts_with")

#> # A tibble: 9 x 16
#>   vernacularName  language update_date acceptedNameUsag… taxonID  scientificName
#>   <chr>           <chr>    <chr>       <chr>             <chr>    <chr>         
#> 1 White-chested … English  2017-03-30  ITIS:1025104      ITIS:10… Cheracebus lu…
#> 2 Widow Monkey    English  2017-03-30  ITIS:1025104      ITIS:10… Cheracebus lu…
#> 3 Medem's Titi    English  2017-03-30  ITIS:1025105      ITIS:10… Cheracebus me…
#> 4 Black-handed T… English  2017-03-30  ITIS:1025105      ITIS:10… Cheracebus me…
#> 5 Collared Titi   English  2017-03-30  ITIS:1025106      ITIS:10… Cheracebus to…
#> 6 White-collared… English  2017-03-30  ITIS:1025106      ITIS:10… Cheracebus to…
#> 7 Yellow-handed … English  2017-03-30  ITIS:1025107      ITIS:10… Cheracebus lu…
#> 8 Rio Purus Titi  English  2017-03-30  ITIS:1025110      ITIS:10… Cheracebus pu…
#> 9 Juruá Collared… English  2017-03-30  ITIS:1025111      ITIS:10… Cheracebus re…
#> # … with 10 more variables: taxonRank <chr>, taxonomicStatus <chr>,
#> #   kingdom <chr>, phylum <chr>, class <chr>, order <chr>, family <chr>,
#> #   genus <chr>, specificEpithet <chr>, infraspecificEpithet <chr>

# }