get_ids

A drop-in replacement for [taxize::get_ids()]

get_ids(names, db = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd"), format = c("bare", "prefix", "uri"), taxadb_db = td_connect(),
  ...)

Arguments

names
a list of scientific names (which may include higher-order ranks in most authorities).

db
abbreviation code for the authority. See details.

format
Format for the returned: bare identifier, one of

bare (e.g. 9606, default, matching taxize::get_ids()),

prefix (e.g. NCBI:9606), or

uri (e.g. http://ncbi.nlm.nih.gov/taxonomy/9606).

taxadb_db
Connection to from [td_connect]().

...
additional arguments passed to ids()

Details

Note that some taxize authorities: nbn, tropicos, and eol, are not recognized by taxadb and will throw an error here. Meanwhile, taxadb recognizes several authorities not known to [taxize::get_ids()]. Both include itis, ncbi, col, and gbif.

Like all taxadb functions, this function will run fastest if a local copy of the authority is installed in advance using [td_create()].

Examples

# NOT RUN {
get_ids("Homo sapiens")
get_ids(c("Homo sapiens", "Mammalia"), format = "prefix")
get_ids("Homo sapiens", db= "ncbi", format = "uri")
# }